+++ services/base-service/src/main/kotlin/dev/redicloud/service/base/events/NodeSuspendedEvent.kt
@@ -1,4 +1,4 @@
-package dev.redicloud.service.node.events
+package dev.redicloud.service.base.events
 
 import dev.redicloud.event.EventFireType
 import dev.redicloud.repository.node.CloudNode
